Moved other opensearch related typed dicts to typing
