+++ sciety_labs/app/routers/api/categorisation/providers.py
@@ -10,7 +10,7 @@ from sciety_labs.app.routers.api.categorisation.typing import (
     ArticleSearchResponseDict,
     CategorisationResponseDict
 )
-from sciety_labs.providers.opensearch.opensearch_article_recommendation import DocumentDict
+from sciety_labs.providers.opensearch.typing import DocumentDict
 from sciety_labs.providers.opensearch.typing import OpenSearchSearchResultDict
