+++ sciety_labs/providers/opensearch/opensearch_article_recommendation.py
@@ -3,7 +3,6 @@ import logging
 from datetime import date, timedelta
 from typing import Any, Iterable, Mapping, Optional, Sequence, cast
 
-from typing_extensions import NotRequired, TypedDict
 
 import numpy.typing as npt
 
@@ -20,6 +19,16 @@ from sciety_labs.providers.interfaces.article_recommendation import (
     SingleArticleRecommendationProvider
 )
 from sciety_labs.providers.crossref import CrossrefMetaDataProvider
+from sciety_labs.providers.opensearch.typing import (
+    DocumentCrossrefAuthorDict,
+    DocumentCrossrefDict,
+    DocumentDict,
+    DocumentEuropePmcAuthorDict,
+    DocumentEuropePmcDict,
+    DocumentS2AuthorDict,
+    DocumentS2Dict,
+    DocumentScietyDict
+)
 from sciety_labs.providers.semantic_scholar import (
     SemanticScholarTitleAbstractEmbeddingVectorProvider
 )
@@ -31,66 +40,6 @@ LOGGER = logging.getLogger(__name__)
 
 
 DEFAULT_OPENSEARCH_MAX_RECOMMENDATIONS = 50
-
-
-class DocumentCrossrefAuthorDict(TypedDict):
-    orcid: NotRequired[str]
-    family_name: NotRequired[str]
-    given_name: NotRequired[str]
-    sequence: NotRequired[str]
-    suffix: NotRequired[str]
-
-
-class DocumentCrossrefDict(TypedDict):
-    title_with_markup: NotRequired[str]
-    publication_date: NotRequired[str]
-    author_list: NotRequired[Sequence[DocumentCrossrefAuthorDict]]
-
-
-class DocumentS2AuthorDict(TypedDict):
-    name: str
-    s2_author_id: NotRequired[str]
-
-
-class DocumentS2Dict(TypedDict):
-    title: str
-    author_list: NotRequired[Sequence[DocumentS2AuthorDict]]
-
-
-class DocumentEuropePmcCollectiveAuthorDict(TypedDict):
-    collective_name: NotRequired[str]
-
-
-class DocumentEuropePmcIndividualAuthorDict(TypedDict):
-    full_name: NotRequired[str]
-    initials: NotRequired[str]
-    last_name: NotRequired[str]
-    first_name: NotRequired[str]
-
-
-class DocumentEuropePmcAuthorDict(
-    DocumentEuropePmcCollectiveAuthorDict,
-    DocumentEuropePmcIndividualAuthorDict
-):
-    pass
-
-
-class DocumentEuropePmcDict(TypedDict):
-    title_with_markup: NotRequired[str]
-    first_publication_date: NotRequired[str]
-    author_list: NotRequired[Sequence[DocumentEuropePmcAuthorDict]]
-
-
-class DocumentScietyDict(TypedDict):
-    evaluation_count: NotRequired[int]
-
-
-class DocumentDict(TypedDict):
-    doi: str
-    crossref: NotRequired[DocumentCrossrefDict]
-    s2: NotRequired[DocumentS2Dict]
-    europepmc: NotRequired[DocumentEuropePmcDict]
-    sciety: NotRequired[DocumentScietyDict]
 
 
 def get_author_names_for_document_s2_authors(
